Add command queue.
`src` commands now are executed sequentially. The upside is that this
works arround current concurrency issues in `src`. The downside is that
now it's possible for certain commands go out of sync with current state
of files, but that shouldn't happen often if everything working as
intended :stuck_out_tongue_closed_eyes:.